+++ src/main/java/it/gov/pagopa/fdr/rest/support/SupportResource.java
@@ -4,7 +4,7 @@ import static it.gov.pagopa.fdr.util.MDCKeys.ACTION;
 import static it.gov.pagopa.fdr.util.MDCKeys.PSP_ID;
 
 import it.gov.pagopa.fdr.rest.support.mapper.SupportResourceServiceMapper;
-import it.gov.pagopa.fdr.rest.support.response.FdrByIuvIurResponse;
+import it.gov.pagopa.fdr.rest.support.response.FdrByPspIdIuvIurResponse;
 import it.gov.pagopa.fdr.service.dto.PaymentGetByPspIdIuvIurDTO;
 import it.gov.pagopa.fdr.service.re.model.FdrActionEnum;
 import it.gov.pagopa.fdr.service.support.FindPaymentsByPspIdAndIuvIurArgs;
@@ -55,7 +55,7 @@ public class SupportResource {
             content =
                 @Content(
                     mediaType = MediaType.APPLICATION_JSON,
-                    schema = @Schema(implementation = FdrByIuvIurResponse.class)))
+                    schema = @Schema(implementation = FdrByPspIdIuvIurResponse.class)))
       })
   @GET
   @Path(
@@ -64,7 +64,7 @@ public class SupportResource {
       + "}/"
   )
   @Re(action = FdrActionEnum.INTERNAL_GET_ALL_BY_PSP_IUV_FDR)
-  public FdrByIuvIurResponse getByIuv(
+  public FdrByPspIdIuvIurResponse getByIuv(
       @PathParam(AppConstant.PSP) @Pattern(regexp = "^(.{1,35})$") String pspId,
       @PathParam(AppConstant.IUV) @Pattern(regexp = "^(.{1,35})$") String iuv,
       @QueryParam(AppConstant.CREATED_FROM) Instant createdFrom,
@@ -86,7 +86,7 @@ public class SupportResource {
                 .createdTo(createdTo)
                 .pageNumber(pageNumber)
                 .pageSize(pageSize).build());
-    return FdrByIuvIurResponse.builder()
+    return FdrByPspIdIuvIurResponse.builder()
         .metadata(mapper.toMetadata(paymentDtoList.getMetadata()))
         .count(paymentDtoList.getCount())
         .data(mapper.toFdrByIuvIurList(paymentDtoList.getData()))
@@ -107,7 +107,7 @@ public class SupportResource {
               content =
               @Content(
                   mediaType = MediaType.APPLICATION_JSON,
-                  schema = @Schema(implementation = FdrByIuvIurResponse.class)))
+                  schema = @Schema(implementation = FdrByPspIdIuvIurResponse.class)))
       })
   @GET
   @Path(
@@ -116,7 +116,7 @@ public class SupportResource {
           + "}/"
   )
   @Re(action = FdrActionEnum.INTERNAL_GET_ALL_BY_PSP_IUR_FDR)
-  public FdrByIuvIurResponse getByIur(
+  public FdrByPspIdIuvIurResponse getByIur(
       @PathParam(AppConstant.PSP) @Pattern(regexp = "^(.{1,35})$") String pspId,
       @PathParam(AppConstant.IUR) @Pattern(regexp = "^(.{1,35})$") String iur,
       @QueryParam(AppConstant.CREATED_FROM) Instant createdFrom,
@@ -138,7 +138,7 @@ public class SupportResource {
                 .createdTo(createdTo)
                 .pageNumber(pageNumber)
                 .pageSize(pageSize).build());
-    return FdrByIuvIurResponse.builder()
+    return FdrByPspIdIuvIurResponse.builder()
         .metadata(mapper.toMetadata(paymentDtoList.getMetadata()))
         .count(paymentDtoList.getCount())
         .data(mapper.toFdrByIuvIurList(paymentDtoList.getData()))
