+++ src/main/java/it/gov/pagopa/fdr/rest/support/mapper/SupportResourceServiceMapper.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.fdr.rest.support.mapper;
 
+import it.gov.pagopa.fdr.rest.model.FdrByPspIdIuvIurBase;
 import it.gov.pagopa.fdr.rest.model.Metadata;
-import it.gov.pagopa.fdr.rest.support.response.FdrByIuvIur;
 import it.gov.pagopa.fdr.service.dto.MetadataDto;
 import it.gov.pagopa.fdr.service.dto.PaymentByPspIdIuvIurDTO;
 import java.util.List;
@@ -13,6 +13,6 @@ import org.mapstruct.factory.Mappers;
 public interface SupportResourceServiceMapper {
   SupportResourceServiceMapper INSTANCE =
       Mappers.getMapper(SupportResourceServiceMapper.class);
-  List<FdrByIuvIur> toFdrByIuvIurList(List<PaymentByPspIdIuvIurDTO> fdrGetPaymentDto);
+  List<FdrByPspIdIuvIurBase> toFdrByIuvIurList(List<PaymentByPspIdIuvIurDTO> fdrGetPaymentDto);
   Metadata toMetadata(MetadataDto metadataDto);
 }
