+++ src/main/java/it/gov/pagopa/fdr/rest/support/response/FdrByPspIdIuvIurResponse.java
@@ -1,6 +1,7 @@
 package it.gov.pagopa.fdr.rest.support.response;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import it.gov.pagopa.fdr.rest.model.FdrByPspIdIuvIurBase;
 import it.gov.pagopa.fdr.rest.model.Metadata;
 import java.util.List;
 import lombok.Getter;
@@ -12,12 +13,12 @@ import org.eclipse.microprofile.openapi.annotations.media.Schema;
 @SuperBuilder
 @Jacksonized
 @JsonPropertyOrder({"metadata", "count", "data"})
-public class FdrByIuvIurResponse {
+public class FdrByPspIdIuvIurResponse {
 
   private Metadata metadata;
 
   @Schema(example = "100")
   private long count;
 
-  private List<FdrByIuvIur> data;
+  private List<FdrByPspIdIuvIurBase> data;
 }
